+++ tests/functions/test_get_primary_keys.py
@@ -3,11 +3,11 @@ try:
 except ImportError:
     from ordereddict import OrderedDict
 import sqlalchemy as sa
-from sqlalchemy_utils import primary_keys
+from sqlalchemy_utils import get_primary_keys
 from tests import TestCase
 
 
-class TestPrimaryKeys(TestCase):
+class TestGetPrimaryKeys(TestCase):
     def create_models(self):
         class Building(self.Base):
             __tablename__ = 'building'
@@ -17,27 +17,28 @@ class TestPrimaryKeys(TestCase):
         self.Building = Building
 
     def test_table(self):
-        assert primary_keys(self.Building.__table__) == OrderedDict({
+        assert get_primary_keys(self.Building.__table__) == OrderedDict({
             '_id': self.Building.__table__.c._id
         })
 
     def test_declarative_class(self):
-        assert primary_keys(self.Building) == OrderedDict({
+        assert get_primary_keys(self.Building) == OrderedDict({
             'id': self.Building.__table__.c._id
         })
 
     def test_declarative_object(self):
-        assert primary_keys(self.Building()) == OrderedDict({
+        assert get_primary_keys(self.Building()) == OrderedDict({
             'id': self.Building.__table__.c._id
         })
 
     def test_class_alias(self):
-        assert primary_keys(sa.orm.aliased(self.Building())) == OrderedDict({
+        alias = sa.orm.aliased(self.Building)
+        assert get_primary_keys(alias) == OrderedDict({
             'id': self.Building.__table__.c._id
         })
 
     def test_table_alias(self):
         alias = sa.orm.aliased(self.Building.__table__)
-        assert primary_keys(alias) == OrderedDict({
+        assert get_primary_keys(alias) == OrderedDict({
             '_id': alias.c._id
         })